Extract general Jujutsu functions out of submit.go
submit.go was getting a little unwieldy.
